backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gevent==23.9.1
greenlet==3.0.1
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
zope.event==5.0
zope.interface==6.1
